requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg2==2.9.9
boto3==1.34.144 
hubspot-api-client==9.0.0
mailersend==0.5.7